requirements.txt
# Core Deep Learning
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
scipy>=1.10.0
pyloudnorm>=0.1.0

# Video/Image Processing
opencv-python>=4.7.0
mediapipe>=0.10.0
numpy>=1.24.0

# Transformers and Pretrained Models
transformers>=4.30.0
timm>=0.9.0

# Data Processing
pandas>=2.0.0
jsonlines>=3.1.0
pyyaml>=6.0
tqdm>=4.65.0
h5py>=3.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Metrics and Calibration
scikit-learn>=1.2.0
netcal>=1.3.0

# Model Optimization
onnx>=1.14.0
onnxruntime>=1.15.0

# Logging and Experiment Tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Type Hints
typing-extensions>=4.5.0

# Utilities
einops>=0.6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

